Convert `Rectangle` from `float` to `int`
This is always used with `int` based `Rectangle` in downstream OPHD. Indeed, it makes the most sense to use `int` values with this class. It uses `Image`, which has `int` based `size` values.

+++ NAS2D/Renderer/RectangleSkin.h
@@ -27,7 +27,7 @@ namespace NAS2D
 
 		Vector<int> minSize() const;
 
-		void draw(Renderer& renderer, const Rectangle<float>& rect) const;
+		void draw(Renderer& renderer, const Rectangle<int>& rect) const;
 
 	private:
 		const Image& mTopLeft;
